+++ src/test/scala/com/github/sudachi0114/logexample/MarkerLoggingSpec.scala
@@ -1,0 +1,11 @@
+package com.github.sudachi0114.logexample
+
+import org.scalatest.funspec.AnyFunSpec
+
+class MarkerLoggingSpec extends AnyFunSpec {
+  describe("logback MarkerLogging test") {
+    it("should call MarkerLogging") {
+      MarkerLogging
+    }
+  }
+}
